[MISC] Transform unary is_function trait to function_like concept.
