+++ test/unit/core/concept/CMakeLists.txt
@@ -1,1 +1,2 @@
 seqan3_test (cereal_test.cpp)
+seqan3_test (concept_function_test.cpp)
